Add autocomplete extension to build
