+++ package/package.json
@@ -15,6 +15,7 @@
     "./copy-button.css": "./dist/copy.css",
     "./code-folding.css": "./dist/folding.css",
     "./scrollbar.css": "./dist/scrollbar.css",
+    "./autocomplete.css": "./dist/autocomplete.css",
     "./setups": "./dist/setups/index.js",
     "./utils": "./dist/utils/index.js",
     "./layout.css": "./dist/layout.css",
@@ -29,6 +30,10 @@
     "./highlight-brackets": "./dist/extensions/matchBrackets/highlight.js",
     "./match-tags": "./dist/extensions/matchTags.js",
     "./tooltips": "./dist/tooltips.js",
+    "./experimental_autocomplete": "./dist/extensions/autocomplete/index.js",
+    "./experimental_autocomplete/css": "./dist/extensions/autocomplete/css/index.js",
+    "./experimental_autocomplete/html": "./dist/extensions/autocomplete/html/index.js",
+    "./experimental_autocomplete/javascript": "./dist/extensions/autocomplete/javascript/index.js",
     "./commands": "./dist/extensions/commands.js",
     "./cursor": "./dist/extensions/cursor.js",
     "./code-folding": "./dist/extensions/folding/index.js",
