+++ package/src/extensions/autocomplete/index.ts
@@ -1,3 +1,5 @@
+/** @module autocomplete */
+
 export { autoComplete, registerCompletions } from "./tooltip.js"
 export { fuzzyFilter, strictFilter } from "./filter.js"
 export * from "./types.js"
