+++ package/vite.config.ts
@@ -19,6 +19,11 @@ const entries = {
 	copy: "src/extensions/copyButton/copy.css",
 	"extensions/copyButton/index": "src/extensions/copyButton/index.ts",
 	"extensions/folding/index": "src/extensions/folding/index.ts",
+	"extensions/autocomplete/index": "src/extensions/autocomplete/index.ts",
+	"extensions/autocomplete/css/index": "src/extensions/autocomplete/css/index.ts",
+	"extensions/autocomplete/html/index": "src/extensions/autocomplete/html/index.ts",
+	"extensions/autocomplete/javascript/index": "src/extensions/autocomplete/javascript/index.ts",
+	autocomplete: "src/extensions/autocomplete/style.css",
 	folding: "src/extensions/folding/folding.css",
 	tooltips: "src/tooltips.ts",
 	scrollbar: "src/scrollbar.css",
